requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
addict=2.4.0=pypi_0
aliyun-python-sdk-core=2.16.0=pypi_0
aliyun-python-sdk-kms=2.16.5=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
asttokens=2.4.1=pypi_0
attrs=24.2.0=pypi_0
blas=1.0=mkl
blinker=1.8.2=pypi_0
brotli-python=1.0.9=py39h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.9.24=h06a4308_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=py39h06a4308_0
cffi=1.17.1=pypi_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=pypi_0
cloudpickle=3.1.0=pypi_0
cmake=3.31.4=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pypi_0
configargparse=1.7=pypi_0
contourpy=1.3.0=pypi_0
crcmod=1.7=pypi_0
cryptography=44.0.0=pypi_0
cubinlinker-cu11=0.3.0.post2=pypi_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-python=11.8.3=pypi_0
cuda-runtime=11.8.0=0
cuda-version=12.6=3
cudf-cu11=24.8.3=pypi_0
cuml-cu11=24.8.0=pypi_0
cupy-cuda11x=13.3.0=pypi_0
cycler=0.12.1=pypi_0
dash=2.18.1=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
dask=2024.7.1=pypi_0
dask-cuda=24.8.2=pypi_0
dask-cudf-cu11=24.8.3=pypi_0
dask-expr=1.1.9=pypi_0
decorator=5.1.1=pypi_0
distributed=2024.7.1=pypi_0
distributed-ucxx-cu11=0.39.1=pypi_0
einops=0.8.0=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.1.0=pypi_0
fastjsonschema=2.20.0=pypi_0
fastrlock=0.8.2=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.14.0=pypi_0
flask=3.0.3=pypi_0
fonttools=4.54.1=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.9.0=pypi_0
ftfy=6.3.1=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
huggingface-hub=0.29.3=pypi_0
idna=3.7=py39h06a4308_0
imageio=2.35.1=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
iopath=0.1.10=pypi_0
ipython=8.18.1=pypi_0
ipywidgets=8.1.5=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.4=py39h06a4308_0
jmespath=0.10.0=pypi_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_3
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter-core=5.7.2=pypi_0
jupyterlab-widgets=3.0.13=pypi_0
kiwisolver=1.4.7=pypi_0
lame=3.100=h7b6447c_0
lazy-loader=0.4=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.11.1.6=0
libcurand=10.3.7.77=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libucx-cu11=1.15.0.post1=pypi_0
libunistring=0.9.10=h27cfd23_0
libwebp-base=1.3.2=h5eee18b_0
lit=18.1.8=pypi_0
llvmlite=0.43.0=pypi_0
locket=1.0.0=pypi_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py39h5eee18b_0
matplotlib=3.9.2=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.10=py39h5eee18b_0
mkl_random=1.2.7=py39h1128e8f_0
mmcv=2.0.0=pypi_0
mmengine=0.10.6=pypi_0
mmsegmentation=1.2.2=pypi_0
model-index=0.1.11=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py39h06a4308_0
msgpack=1.1.0=pypi_0
multiscaledeformableattention=1.0=pypi_0
mypy-extensions=1.0.0=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=py39h06a4308_0
numba=0.60.0=pypi_0
numpy=1.26.4=pypi_0
nvtx=0.2.10=pypi_0
omegaconf=2.3.0=pypi_0
open3d=0.18.0=pypi_0
opencv-python=4.10.0.84=pypi_0
opendatalab=0.0.10=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openmim=0.3.9=pypi_0
openssl=3.0.15=h5eee18b_0
openxlab=0.1.2=pypi_0
ordered-set=4.1.0=pypi_0
oss2=2.17.0=pypi_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
parso=0.8.4=pypi_0
partd=1.4.2=pypi_0
pexpect=4.9.0=pypi_0
pillow=10.4.0=py39h5eee18b_0
pip=24.2=py39h06a4308_0
platformdirs=4.3.6=pypi_0
plotly=5.24.1=pypi_0
portalocker=2.10.1=pypi_0
prettytable=3.13.0=pypi_0
prompt-toolkit=3.0.48=pypi_0
protobuf=5.28.2=pypi_0
psutil=6.0.0=pypi_0
ptxcompiler-cu11=0.8.1.post2=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pyarrow=16.1.0=pypi_0
pycparser=2.22=pypi_0
pycryptodome=3.21.0=pypi_0
pygments=2.18.0=pypi_0
pylibraft-cu11=24.8.1=pypi_0
pynvml=11.4.1=pypi_0
pyparsing=3.1.4=pypi_0
pyquaternion=0.9.9=pypi_0
pyre-extensions=0.0.23=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.20=he870216_1
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.0.0=py3.9_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2023.4=pypi_0
pyyaml=6.0.2=pypi_0
raft-dask-cu11=24.8.1=pypi_0
rapids-dask-dependency=24.8.0=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.11.6=pypi_0
requests=2.28.2=pypi_0
retrying=1.3.4=pypi_0
rich=13.4.2=pypi_0
rmm-cu11=24.8.2=pypi_0
rpds-py=0.20.0=pypi_0
safetensors=0.4.5=pypi_0
scikit-image=0.24.0=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.13.1=pypi_0
seaborn=0.13.2=pypi_0
setuptools=60.2.0=pypi_0
six=1.16.0=pypi_0
sortedcontainers=2.4.0=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
submitit=1.5.2=pypi_0
sympy=1.13.2=py39h06a4308_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tblib=3.0.0=pypi_0
tenacity=9.0.0=pypi_0
tensorboardx=2.6.2.2=pypi_0
termcolor=2.5.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.8.30=pypi_0
timm=1.0.9=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.1=pypi_0
tomli=2.2.1=pypi_0
toolz=1.0.0=pypi_0
torchaudio=2.0.0=py39_cu118
torchmetrics=0.10.3=pypi_0
torchtriton=2.0.0=py39
torchvision=0.15.0=py39_cu118
tornado=6.4.1=pypi_0
tqdm=4.65.2=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.49.0=pypi_0
treelite=4.3.0=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.11.0=py39h06a4308_0
tzdata=2024.2=pypi_0
ucx-py-cu11=0.39.2=pypi_0
ucxx-cu11=0.39.1=pypi_0
urllib3=1.26.20=pypi_0
wcwidth=0.2.13=pypi_0
websocket-client=1.8.0=pypi_0
werkzeug=3.0.4=pypi_0
wheel=0.44.0=py39h06a4308_0
widgetsnbextension=4.0.13=pypi_0
xformers=0.0.18=pypi_0
xz=5.4.6=h5eee18b_1
yacs=0.1.8=pypi_0
yapf=0.43.0=pypi_0
zict=3.0.0=pypi_0
zipp=3.20.2=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
